Edit recall.html
Add a context to radio buttons

As it was just a bunch of random text so adding proper names and adding a
heading  made it much more sensible.

Change top margin of nav element in stylesheet

As there was a white line visible between the h1 element and nav, shifting
nav 1px top fixed the issue.

Change random ordered list to login instruction

To make the document go with the rest of the changes this change was made.